authorized requests code added

+++ front-end/src/components/WorkoutDetails.js
@@ -1,4 +1,5 @@
 import { useWorkoutsContext } from "../context/hooks/useWorkoutsContext";
+import { useAuthContext } from "../context/hooks/useAuthContext";
 
 // date fns
 
@@ -6,10 +7,16 @@ import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch("/api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: { Authorization: `Bearer ${user.token}` },
     });
     const json = await response.json();
